requirements.txt
sentence-transformers
pandas
numpy
scikit-learn
PyYAML
python-Levenshtein